ai-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
diffusers==0.25.0
transformers==4.36.2
accelerate==0.25.0
xformers==0.0.23
controlnet-aux==0.4.0
opencv-python==4.8.1.78
numpy==1.25.2
scipy==1.11.4
safetensors==0.4.1
huggingface-hub==0.19.4
datasets==2.15.0
peft==0.7.1
bitsandbytes==0.41.3
compel==2.0.2
# ControlNet models for SDXL
diffusers[controlnet]==0.25.0